+++ tests/testthat/test-tab_afun_cfun.R
@@ -185,6 +185,61 @@ test_that("Error localization for missing split variable when done in alt_count_
     expect_silent(lyt_row %>% build_table(ex_adsl, alt_counts_df = DM))
 })
 
+test_that("Error localization for missmatch split variable when done in alt_count_df", {
+    afun_tmp <- function(x, .alt_df_row, .spl_context,...) {
+        # Important check that order is aligned even if source levels are not
+        check_val <- unique(.alt_df_row$ARMCD)
+        # This is something mysterious happening in splits for which if the values are all
+        # NAs in the split column, the dataspl has the nrow of the data in NA rows. xxx ToFix
+        check_val <- check_val[!is.na(check_val)] 
+        stopifnot(as.character(check_val) == .spl_context$value[2])
+        mean(x)
+    }
+    lyt_row <- basic_table() %>% split_rows_by("ARMCD") %>% analyze("BMRKR1", afun = afun_tmp)
+    
+    # Mismatch in the number of splits (NA is 0)
+    DM_tmp <- DM %>% mutate("ARMCD" = NA_character_)
+    expect_error(lyt_row %>% build_table(ex_adsl, alt_counts_df = DM_tmp), 
+                 regexp = paste0("alt_counts_df split variable\\(s\\) \\[ARMCD\\] *"))
+    
+    # Mismatch of levels
+    armcd_col <- factor(sample(c("arm A", "arm B", "arm C"), nrow(DM), replace = TRUE))
+    DM_tmp <- DM %>% mutate("ARMCD" = armcd_col)
+    expect_error(lyt_row %>% build_table(ex_adsl, alt_counts_df = DM_tmp), 
+                 regexp = paste0("alt_counts_df split variable\\(s\\) \\[ARMCD\\] *"))
+    
+    # Mix mismatch of levels
+    armcd_col <- factor(sample(c("arm A", "ARM B", "ARM C"), nrow(DM), replace = TRUE))
+    DM_tmp <- DM %>% mutate("ARMCD" = armcd_col)
+    expect_error(lyt_row %>% build_table(ex_adsl, alt_counts_df = DM_tmp), 
+                 regexp = paste0("alt_counts_df split variable\\(s\\) \\[ARMCD\\] *"))
+    
+    # Mismatch in the number of levels 
+    armcd_col2 <- factor(sample(levels(ex_adsl$ARMCD)[c(1, 2)], nrow(DM), replace = TRUE))
+    DM_tmp <- DM %>% mutate("ARMCD" = armcd_col2)
+    expect_error(lyt_row %>% build_table(ex_adsl, alt_counts_df = DM_tmp), 
+                 regexp = paste0("alt_counts_df split variable\\(s\\) \\[ARMCD\\] *"))
+    
+    # Another order -> should work? yes, split is valid
+    levels(armcd_col) <- levels(ex_adsl$ARMCD)[c(1, 3, 2)]
+    DM_tmp <- DM %>% mutate("ARMCD" = armcd_col)
+    expect_silent(lyt_row %>% build_table(ex_adsl, alt_counts_df = DM_tmp))
+    
+    # Mix mismatch of levels but covering them all -> valid split
+    armcd_col <- factor(sample(c("arm A", levels(ex_adsl$ARMCD)), nrow(DM), replace = TRUE))
+    DM_tmp <- DM %>% mutate("ARMCD" = armcd_col)
+    expect_silent(lyt_row %>% build_table(ex_adsl, alt_counts_df = DM_tmp))
+    
+    # Values are all NA, but the levels are correct
+    DM_tmp$ARMCD <- factor(NA, levels = levels(ex_adsl$ARMCD))
+    expect_error(lyt_row %>% build_table(ex_adsl, alt_counts_df = DM_tmp), 
+                 regexp = paste0("alt_counts_df split variable\\(s\\) \\[ARMCD\\] *"))
+    
+    DM_tmp$ARMCD <- factor(NA, levels = levels(ex_adsl$ARMCD))
+    DM_tmp$ARMCD[seq_along(levels(ex_adsl$ARMCD))] <- levels(ex_adsl$ARMCD) 
+    expect_silent(lyt_row %>% build_table(ex_adsl, alt_counts_df = DM_tmp))
+})
+
 context("Content functions (cfun)")
 
 test_that(".alt_df_row appears in cfun but not in afun.", {