+++ NEWS.md
@@ -16,6 +16,7 @@
  * Cleaned up spelling in documentation ([#685](https://github.com/insightsengineering/rtables/issues/685)) 
  * Added `qtable_layout` and fixed `qtable` labeling via `row_labels` ([#698](https://github.com/insightsengineering/rtables/issues/698))
  * Added vignette on exploratory analysis with `qtable`.
+ * Error catching and test coverage for cases where `alt_counts_df` presents different splits from `df`.
 
 ## rtables 0.6.2
  * Fixed major regressions for `page_by` machinery caused by migration to `formatters` 0.5.1 pagination framework.